Task 2/requirements.txt
torch==2.4.0
torchvision==0.19.0
numpy==1.24.3
matplotlib==3.7.1
Pillow==9.4.0
scikit-learn==1.2.2
opencv-python==4.8.0.74
